[Fix] : [Chore] : Tuist side effect issue
Git : #31
