+++ Projects/Feature/AD_Feature/Sources/Extensions/Extension+Moya.swift
@@ -1,14 +1,14 @@
 //
 //  Extension+Moya.swift
-//  AD_Utils
+//  AD_Feature
 //
-//  Created by minii on 2023/06/30.
+//  Created by minii on 2023/09/04.
 //  Copyright © 2023 chminipark. All rights reserved.
 //
 
 import Moya
 
-public extension MoyaProvider {
+extension MoyaProvider {
   func request(_ target: Target) async -> Result<Response, MoyaError> {
     await withCheckedContinuation { continuation in
       self.request(target) { result in
