+++ Projects/Feature/AD_Feature/Sources/Types/ADError/ADMoyaError.swift
@@ -1,0 +1,16 @@
+//
+//  ADMoyaError.swift
+//  AD_Feature
+//
+//  Created by minii on 2023/08/07.
+//  Copyright © 2023 chminipark. All rights reserved.
+//
+
+import Foundation
+
+public enum ADMoyaError: Error {
+  case jsonMapping
+  case calculateInServer
+  case connection
+  case imageMapping
+}
