+++ Projects/Share/AD_Utils/Project.swift
@@ -5,9 +5,6 @@ let project = Project.makeModule(
   myModule: .AD_Utils,
   platform: .iOS,
   product: .framework,
-  dependencies: [
-    .ComposableArchitecture
-  ],
   resources: ["Resources/**"],
   withTest: false
 )
