+++ Projects/UI/AD_UI/Project.swift
@@ -4,7 +4,6 @@ import ProjectEnvironment
 let project = Project.makeModule(
   myModule: .AD_UI,
   platform: .iOS,
-//  product: .staticFramework,
   product: .framework,
   dependencies: [
     .AD_CropImage,
