+++ Projects/UI/AD_UI/Sources/Protocol/ADUI.swift
@@ -1,15 +1,15 @@
 //
 //  ADUI.swift
-//  AD_Utils
+//  AD_UI
 //
-//  Created by minii on 2023/06/09.
+//  Created by minii on 2023/09/04.
 //  Copyright © 2023 chminipark. All rights reserved.
 //
 
 import SwiftUI
 import ComposableArchitecture
 
-public protocol ADUI: View {
+protocol ADUI: View {
   associatedtype MyFeature: Reducer
   associatedtype MyViewStore = ViewStore<MyFeature.State, MyFeature.Action>
