+++ Projects/UI/AD_UIHelper/AD_CropImage/Project.swift
@@ -11,7 +11,6 @@ import ProjectEnvironment
 let project = Project.makeModule(
   myModule: .AD_CropImage,
   platform: .iOS,
-//  product: .staticFramework,
   product: .framework,
   dependencies: [
     .AD_Utils
