+++ Projects/UI/AD_UIHelper/AD_MaskingImage/Project.swift
@@ -11,7 +11,6 @@ import ProjectEnvironment
 let project = Project.makeModule(
   myModule: .AD_MaskingImage,
   platform: .iOS,
-//  product: .staticFramework,
   product: .framework,
   dependencies: [
     .AD_Utils
